// archive_v1_20250930_214450/src/types/dify-workflow.ts
export interface DifyApp {
  description?: string
  icon: string
  icon_background: string
  mode: 'workflow' | 'advanced-chat' | 'agent-chat' | 'chat'
  name: string
  use_icon_as_answer_icon?: boolean
}

export interface DifyEnvironmentVariable {
  variable: string
  name: string
  type?: 'string' | 'number' | 'boolean' | 'secret'
  value: string
}

export interface DifyFeatures {
  file_upload?: {
    enabled?: boolean
    image?: {
      enabled: boolean
      number_limits: number
      transfer_methods: ('local_file' | 'remote_url')[]
      max_size_mb?: number
    }
    number_limits?: number
    max_size_mb?: number
  }
  opening_statement?: string
  retriever_resource?: {
    enabled: boolean
    top_k?: number
  }
  sensitive_word_avoidance?: {
    enabled: boolean
    type: 'moderation'
  }
  speech_to_text?: {
    enabled: boolean
    provider: string
  }
  text_to_speech?: {
    enabled: boolean
    provider: string
    voice?: string
  }
  suggested_questions?: string[]
}
// ...
export interface DifyWorkflowGraph {
  edges: DifyEdge[]
  nodes: DifyNode[]
  viewport?: DifyViewport
}

export interface DifyWorkflow {
  conversation_variables?: unknown[]
  environment_variables: DifyEnvironmentVariable[]
  features: DifyFeatures
  graph: DifyWorkflowGraph
}

export interface DifyDSLFile {
  app: DifyApp
  kind: 'app'
  version: string
  workflow: DifyWorkflow
}